refactor(chat): address review — safe tab focus, lightweight memory diff, doc/path fixes
