+++ apps/docs/content/docs/skills.mdx
@@ -57,7 +57,7 @@ The Skills tab's "Try in chat" button on each skill prefills the side panel chat
 
 ## Tools the agent uses
 
-The skill subsystem is exposed to the agent through four tools:
+The skill subsystem is exposed to the agent through three tools:
 
 | Tool | Description |
 | --- | --- |
