+++ apps/extension/src/components/chat/tool-results/memory.tsx
@@ -1,5 +1,6 @@
 import { Database } from "lucide-react";
 import { HighlightedCode } from "./highlighted-code";
+import { buildAdditionDiff } from "@/lib/agent/tools/memory-diff";
 
 interface Props {
   args: Record<string, unknown>;
@@ -7,68 +8,12 @@ interface Props {
   action: "save" | "update";
 }
 
-/**
- * Build a unified-diff-style string (lines prefixed with ` `, `+`, or `-`)
- * from old/new content using a longest-common-subsequence line diff. Unchanged
- * lines render as context; only genuine changes get +/- markers. Exported for
- * unit testing.
- */
-export function buildMemoryDiff(oldContent: string, newContent: string): string {
-  const a = oldContent.length ? oldContent.split("\n") : [];
-  const b = newContent.length ? newContent.split("\n") : [];
-
-  // LCS table (classic dynamic-programming line diff).
-  const m = a.length;
-  const n = b.length;
-  const lcs: number[][] = Array.from({ length: m + 1 }, () =>
-    new Array<number>(n + 1).fill(0),
-  );
-  for (let i = m - 1; i >= 0; i--) {
-    for (let j = n - 1; j >= 0; j--) {
-      lcs[i][j] =
-        a[i] === b[j]
-          ? lcs[i + 1][j + 1] + 1
-          : Math.max(lcs[i + 1][j], lcs[i][j + 1]);
-    }
-  }
-
-  const lines: string[] = [];
-  let i = 0;
-  let j = 0;
-  while (i < m && j < n) {
-    if (a[i] === b[j]) {
-      lines.push(`  ${a[i]}`);
-      i++;
-      j++;
-    } else if (lcs[i + 1][j] >= lcs[i][j + 1]) {
-      lines.push(`- ${a[i]}`);
-      i++;
-    } else {
-      lines.push(`+ ${b[j]}`);
-      j++;
-    }
-  }
-  while (i < m) lines.push(`- ${a[i++]}`);
-  while (j < n) lines.push(`+ ${b[j++]}`);
-
-  return lines.join("\n");
-}
-
-/** All-additions diff for a brand-new memory. */
-export function buildAdditionDiff(content: string): string {
-  if (!content) return "";
-  return content
-    .split("\n")
-    .map((l) => `+ ${l}`)
-    .join("\n");
-}
-
 export function MemoryResult({ args, result, action }: Props) {
   const content = typeof args.content === "string" ? args.content : "";
 
   // Handle the "didn't happen" outcomes (e.g. update failed, save collided).
   const r = result as
-    | { saved?: boolean; updated?: boolean; reason?: string; oldContent?: string; newContent?: string }
+    | { saved?: boolean; updated?: boolean; reason?: string; diffPreview?: string }
     | undefined;
   const failed =
     r != null &&
@@ -76,8 +21,10 @@ export function MemoryResult({ args, result, action }: Props) {
       (action === "update" && r.updated === false));
 
   let diffText: string;
-  if (action === "update" && typeof r?.oldContent === "string" && typeof r?.newContent === "string") {
-    diffText = buildMemoryDiff(r.oldContent, r.newContent);
+  if (action === "update" && typeof r?.diffPreview === "string") {
+    // The updateMemory tool computes the diff at execute time (keeps the
+    // result lightweight — no full memory bodies in the transcript).
+    diffText = r.diffPreview;
   } else {
     // Save (or update before result resolves): show the content as additions.
     diffText = buildAdditionDiff(content);
